+++ router/CHANGELOG.md
@@ -4,6 +4,55 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.119.1](https://github.com/wundergraph/cosmo/compare/router@0.119.0...router@0.119.1) (2024-10-03)
+
+### Bug Fixes
+
+* apollo compatibility for truncating floats, invalid __typenames ([#1242](https://github.com/wundergraph/cosmo/issues/1242)) ([a773ea5](https://github.com/wundergraph/cosmo/commit/a773ea53d7f9abe48033dad822f4424a31e28af2)) (@devsergiy)
+
+# [0.119.0](https://github.com/wundergraph/cosmo/compare/router@0.118.0...router@0.119.0) (2024-10-03)
+
+### Bug Fixes
+
+* fix panic when accessController is nil in handleUpgradeRequest ([#1239](https://github.com/wundergraph/cosmo/issues/1239)) ([5699c03](https://github.com/wundergraph/cosmo/commit/5699c03db4d0b13b39643787700e49a389cf038d)) (@ElliottZeroFlucs)
+
+### Features
+
+* add option to add response header with trace id ([#1234](https://github.com/wundergraph/cosmo/issues/1234)) ([f8f5078](https://github.com/wundergraph/cosmo/commit/f8f50781917ee4bb0c24ac0c8aa9db71f87e9a05)) (@JivusAyrus)
+* add option to have custom names client name and version headers ([#1233](https://github.com/wundergraph/cosmo/issues/1233)) ([e09348d](https://github.com/wundergraph/cosmo/commit/e09348d2925469e259b526c231a14c6abcc80916)) (@JivusAyrus)
+* router version command ([#1240](https://github.com/wundergraph/cosmo/issues/1240)) ([39970e5](https://github.com/wundergraph/cosmo/commit/39970e5b1c6d8e2a2958f5ea55bb0d8aa7206d92)) (@StarpTech)
+
+# [0.118.0](https://github.com/wundergraph/cosmo/compare/router@0.117.0...router@0.118.0) (2024-10-03)
+
+### Features
+
+* advanced access logs ([#1203](https://github.com/wundergraph/cosmo/issues/1203)) ([b4dc9ac](https://github.com/wundergraph/cosmo/commit/b4dc9acd964b2982fa9e27ec066e91583991ee17)) (@StarpTech)
+
+# [0.117.0](https://github.com/wundergraph/cosmo/compare/router@0.116.3...router@0.117.0) (2024-10-02)
+
+### Features
+
+* add apollo compatibility flag to truncate floats ([#1236](https://github.com/wundergraph/cosmo/issues/1236)) ([17c80c9](https://github.com/wundergraph/cosmo/commit/17c80c91f4e594cf48dae887840517b588bbe0f8)) (@jensneuse)
+
+## [0.116.3](https://github.com/wundergraph/cosmo/compare/router@0.116.2...router@0.116.3) (2024-09-30)
+
+### Bug Fixes
+
+* upgrade engine for ws deadlock, remove pont pool ([#1230](https://github.com/wundergraph/cosmo/issues/1230)) ([3ac50f8](https://github.com/wundergraph/cosmo/commit/3ac50f8b878fb6a5e2460c22aa94412a28c49600)) (@StarpTech)
+
+## [0.116.2](https://github.com/wundergraph/cosmo/compare/router@0.116.1...router@0.116.2) (2024-09-30)
+
+### Bug Fixes
+
+* value completion reset, preserve __typename field location ([#1228](https://github.com/wundergraph/cosmo/issues/1228)) ([9d671f0](https://github.com/wundergraph/cosmo/commit/9d671f071af6d6b5789968682b98a9c5b7f9a7ec)) (@devsergiy)
+* **websocket:** avoid data race on client request ([#1226](https://github.com/wundergraph/cosmo/issues/1226)) ([dd3ef36](https://github.com/wundergraph/cosmo/commit/dd3ef360447afae8c95149ac71f4f9e1c15f66dd)) (@StarpTech)
+
+## [0.116.1](https://github.com/wundergraph/cosmo/compare/router@0.116.0...router@0.116.1) (2024-09-30)
+
+### Bug Fixes
+
+* planning of consecutive fragments and fragments on union ([#1223](https://github.com/wundergraph/cosmo/issues/1223)) ([3e390cd](https://github.com/wundergraph/cosmo/commit/3e390cd5cf6ca87694c6a25794aecd84d4f0c31e)) (@devsergiy)
+
 # [0.116.0](https://github.com/wundergraph/cosmo/compare/router@0.115.1...router@0.116.0) (2024-09-27)
 
 ### Features